Finish implementing the event, emitter, multi emitter, and tests
