+++ Gemfile
@@ -4,6 +4,8 @@ source 'https://rubygems.org'
 
 gemspec
 
+gem 'active_record_value_objects', github: 'joegaudet/active_record_value_objects'
+
 gem 'minitest', '~> 5.25'
 gem 'rubocop'
 gem 'rubocop-minitest'
