--- lib/sidekiq/events/errors/no_handler_error.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-module Sidekiq
-  module Events
-    module Errors
-      class NoHandlerError < NoMethodError; end
-    end
-  end
-end
