+++ lib/sidekiq_events.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+require 'sidekiq_events/configuration'
+require 'sidekiq_events/emitter'
+require 'sidekiq_events/event'
+require 'sidekiq_events/multi_emitter'
+require 'sidekiq_events/errors/no_handler_error'
+
+module SidekiqEvents
+end
