+++ lib/sidekiq_events/errors/no_handler_error.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module SidekiqEvents
+  module Errors
+    class NoHandlerError < NoMethodError; end
+  end
+end
