+++ test/sidekiq_events/emitter_test.rb
@@ -1,28 +1,24 @@
 # frozen_string_literal: true
 
 require 'minitest/autorun'
-require 'sidekiq/events/event'
-require 'sidekiq/events/emitter'
+require 'sidekiq_events'
 
-class MockEmitter < Sidekiq::Events::Emitter
+class MockEmitter < SidekiqEvents::Emitter
   # Used to test if the _event_source is set to the inherited class, or this one
 end
 
-class MockEmitterEvent < Sidekiq::Events::Event
+class MockEmitterEvent < SidekiqEvents::Event
   event_name :bar
 
-  def foo
-    'foo'
-  end
+  attribute :foo, Types::String.optional
 end
 
-describe Sidekiq::Events::Emitter do
+describe SidekiqEvents::Emitter do
   it 'extracts event name, attributes, and event source class from an event and publishes them over wisper' do
     event = MockEmitterEvent.new(foo: 'test')
     emitter = MockEmitter.new
 
     assert(emitter.call(event))
-    assert_equal({ foo: 'test', _id: event._id, emitted_at: event.emitted_at, _event_source: 'MockEmitter' },
-                 emitter.attrs)
+    assert_equal({ foo: 'test', _id: event._id, emitted_at: event.emitted_at, _event_class: event.class.name, _event_source: 'MockEmitter' }, emitter.attributes)
   end
 end
