+++ test/sidekiq_events/event_test.rb
@@ -1,39 +1,38 @@
 # frozen_string_literal: true
 
 require 'minitest/autorun'
-require 'sidekiq/events/event'
+require 'sidekiq_events'
 
-class MockEmitterEvent < Sidekiq::Events::Event
+class MockEvent < SidekiqEvents::Event
   event_name 'my_custom_event'
 
   sidekiq_options queue: 'default', retry: 3
 
-  def foo
-    'foo'
-  end
+  attribute :foo, Types::String.optional
 end
 
-describe Sidekiq::Events::Event do
+describe SidekiqEvents::Event do
   it 'correctly set the attributes using .new' do
-    event = MockEmitterEvent.new(foo: 'test')
+    event = MockEvent.new(foo: 'test')
 
-    assert_equal({ foo: 'test', _id: event._id, emitted_at: event.emitted_at }, event.attributes)
+    # Emitted At and Event Source are set when the event is actually emitted by the emitter
+    assert_equal({ foo: 'test', _id: event._id, emitted_at: nil, _event_source: nil }, event.attributes)
   end
 
   it 'correctly set the channels' do
-    event = MockEmitterEvent.new(foo: 'test')
+    event = MockEvent.new(foo: 'test')
 
     assert_equal(['my_custom_event'], event.channels)
   end
 
   it 'correctly set the event_name' do
-    event = MockEmitterEvent.new
+    event = MockEvent.new
 
     assert_equal('my_custom_event', event.event_name)
   end
 
   it 'correctly set the sidekiq_options' do
-    event = MockEmitterEvent.new
+    event = MockEvent.new
 
     assert_equal({ queue: 'default', retry: 3 }, event.sidekiq_options)
   end
